+++ public/html/board-write.html
@@ -48,7 +48,7 @@
             </article> 
 
             <article class="main">
-                <form action="http://localhost:3001/boards" method="post">
+                <form method="post">
                     <label for="postTitle"><h3>제목 * </h3></label>
                     <hr class="horizontal-rule"/> 
                     <input type="text" name="postTitle" id = "title" placeholder="제목을 입력해주세요. (최대 26글자)">
